1.添加MyBatisPlusConfig 2.添加PageUtil 3.更改MyBatisPlus版本号 4.添加组别表分页查询

+++ lk-web-common/src/main/java/com/lucky/kali/common/base/BaseService.java
@@ -123,7 +123,7 @@ public interface BaseService<E extends BaseEntity, D extends BaseDTO> {
      *
      * @param queryWrapper 实体对象封装操作类（可以为 null）
      */
-    Integer selectCount(@Param(Constants.WRAPPER) Wrapper<E> queryWrapper);
+    Long selectCount(@Param(Constants.WRAPPER) Wrapper<E> queryWrapper);
 
     /**
      * 根据 entity 条件，查询全部记录
